feat: Enhance forecast charts with gradients and improved UI
- Updated HourlyForecastChart and DailyForecastChart to use gradient colors for temperature and precipitation bars.
- Improved text styles and layout for temperature and precipitation probability labels.
- Added icons to enhance visual representation in charts.
- Adjusted padding and dimensions for better spacing and aesthetics.

feat: Revamp HourlyForecastList with a modern design

- Redesigned HourlyForecastList to include a blurred background effect and rounded corners.
- Enhanced the display of weather details with improved typography and layout.
- Added a new method to handle wind unit conversion based on user settings.

feat: Implement auto-refresh for weather data

- Added a timer to automatically refresh weather and air quality data every 5 minutes.
- Integrated refresh functionality into the HomePage with a pull-to-refresh feature.

fix: Update UV index handling in health cards

- Changed UV index type from int to double for more accurate representation.
- Updated related methods to handle double values and display UV index with one decimal precision.

feat: Introduce background fetch for periodic weather updates

- Created BackgroundFetchService to manage background tasks for weather updates.
- Configured background fetch to retrieve weather data based on the last known location.

feat: Add UV index response model for API integration

- Implemented UVIndexResponse model to handle UV index data from the API.
- Generated JSON serialization methods for the UV index model.

+++ lib/domain/entities/weather.dart
@@ -16,7 +16,7 @@ abstract class Weather with _$Weather {
     required int pressure,
     required int visibility,
     required double dewPoint,
-    required int uvIndex,
+    required double uvIndex,
     required DateTime sunrise,
     required DateTime sunset,
     required DateTime timestamp,
